Support dbt 1 0 [sc-5881]
