+++ tests/dbt/test_extractor.py
@@ -38,6 +38,15 @@ async def test_trial_project_v3(test_root_dir):
 
 
 @pytest.mark.asyncio
+async def test_trial_project_v4(test_root_dir):
+    await _test_project(
+        test_root_dir + "/dbt/data/trial_v4",
+        "http://localhost:8080",
+        "https://github.com/MetaphorData/dbt/tree/main/trial",
+    )
+
+
+@pytest.mark.asyncio
 async def test_ride_share_project(test_root_dir):
     await _test_project(test_root_dir + "/dbt/data/ride_share")
